requirement.txt
absl-py==1.2.0
addict==2.4.0
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
codecov==2.1.12
colorama==0.4.6
coverage==7.1.0
cycler==0.11.0
Cython==0.29.33
docker-pycreds==0.4.0
dominate==2.7.0
exceptiongroup==1.1.0
face-alignment==1.3.5
filelock==3.9.0
flake8==5.0.4
flatbuffers==2.0
fonttools==4.34.4
gast==0.5.3
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
idna==3.3
imageio==2.20.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.2.0
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.6
libclang==14.0.1
llvmlite==0.39.0
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
mkl-fft==1.3.1
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.4.2
mypy-extensions==0.4.3
networkx==2.6.3
ninja==1.11.1
numba==0.56.0
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pathtools==0.1.2
Pillow==9.2.0
pluggy==1.0.0
prettytable==3.6.0
progress==1.6
protobuf==3.20.1
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pyrallis==0.3.1
PySocks==1.7.1
pytest==7.2.1
pytest-runner==6.0.0
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resample2d-cuda==0.0.0
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
sentry-sdk==1.16.0
setproctitle==1.3.2
smmap==5.0.0
soupsieve==2.4
tabulate==0.9.0
tdqm==0.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.1
torch==1.12.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.13.0
tqdm==4.64.0
trimesh==3.12.9
typing-inspect==0.8.0
urllib3==1.26.11
wandb==0.13.10
wcwidth==0.2.6
Werkzeug==2.2.1
wget==3.2
wrapt==1.14.1
yapf==0.32.0
zipp==3.8.1
